kill works

+++ Midterm/server.c
@@ -8,18 +8,41 @@
 #include <errno.h> 
 #include <string.h> 
 
-volatile sig_atomic_t kill_signal_received = 0;
+int kill_signal_received = 0;
 
 pthread_mutex_t fd_mutex = PTHREAD_MUTEX_INITIALIZER;
-pthread_mutex_t global_var_mutex = PTHREAD_MUTEX_INITIALIZER;
 pthread_mutex_t waiting_list_mutex = PTHREAD_MUTEX_INITIALIZER;
 pthread_mutex_t connected_list_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 void sigint_handler(int sig) 
 {
-    pthread_mutex_lock(&global_var_mutex);
     kill_signal_received = 1;
-    pthread_mutex_unlock(&global_var_mutex);
+    exit(EXIT_SUCCESS);
+}
+
+void kill_parent_and_children(pid_t parent_pid) 
+{
+    // Kill the parent process
+    kill(parent_pid, SIGKILL);
+
+    // Wait for the parent process to terminate
+    waitpid(parent_pid, NULL, 0);
+
+    // Kill all child processes recursively
+    pid_t child_pid;
+    while ((child_pid = fork()) != 0) {
+        if (child_pid == -1) {
+            perror("Error forking child process");
+            exit(EXIT_FAILURE);
+        } else if (child_pid == 0) {
+            // Child process
+            kill(getpid(), SIGKILL); // Kill itself
+            exit(EXIT_SUCCESS);
+        } else {
+            // Parent process
+            waitpid(child_pid, NULL, 0); // Wait for child to terminate
+        }
+    }
 }
 
 char* help_available_operations()
@@ -28,45 +51,54 @@ char* help_available_operations()
            "help, list, readF, writeT, upload, download, archServer, quit, killServer \n";
 }
 
-char* help_for_operation(operation_type_t operation)
+char* help_for_operation(char* commandAsked)
 {
-    switch(operation)
+    if(strcmp(commandAsked, "help") == 0)
     {
-        case HELP:
-            return "Display the list of possible client requests.\n";
-
-        case LIST:
-            return "Sends a request to display the list of files in Server's directory. (Also displays the list received from the Server)\n";
-
-        case READ_FILE:
-            return "readF <file> <line #> \n"
-                     "requests to display the # line of the <file>, if no line number is given the whole contents of the file is requested (and displayed on the client side) \n";
-
-        case WRITE_FILE:
-            return "writeT <file> <line #> <string>\n"
-                    "Request to write the content of “string” to the #th line the <file>, if the line # is not given writes to the end of file. If the file does not exists in Servers directory creates and edits the file at the same time.\n";
-
-        case UPLOAD:
-            return "upload <file>\n"
-                    "Uploads the file from the current working directory of client to the Server's directory. (Beware of the cases no file in clients current working directory  and  file with the same name on Server's side)\n";
-
-        case DOWNLOAD:
-            return "download <file>\n"
-                    "Request to receive <file> from Server's directory to client side.\n";
-
-        case ARCHIVE_SERVER:
-            return "archServer <fileName>.tar\n"
-                    "Using fork, exec and tar utilities create a child process that will collect all the files currently available on the the Server side and store them in the <filename>.tar archive.\n";
-
-        case KILL_SERVER:
-            return "killServer\n"
-                    "Sends a kill request to the Server\n";
-
-        case QUIT:
-            return "quit\n"
-                    "Send write request to Server side log file and quits.\n";
-        default:
-            return "Invalid operation\n";
+        return "Display the list of possible client requests.\n";
+    }
+    else if(strcmp(commandAsked, "list") == 0)
+    {
+        return "Sends a request to display the list of files in Server's directory. (Also displays the list received from the Server)\n";
+    }
+    else if(strcmp(commandAsked, "readF") == 0)
+    {
+        return "readF <file> <line #> \n"
+               "requests to display the # line of the <file>, if no line number is given the whole contents of the file is requested (and displayed on the client side) \n";
+    }
+    else if(strcmp(commandAsked, "writeT") == 0)
+    {
+        return "writeT <file> <line #> <string>\n"
+               "Request to write the content of “string” to the #th line the <file>, if the line # is not given writes to the end of file. If the file does not exists in Servers directory creates and edits the file at the same time.\n";
+    }
+    else if(strcmp(commandAsked, "upload") == 0)
+    {
+        return "upload <file>\n"
+               "Uploads the file from the current working directory of client to the Server's directory. (Beware of the cases no file in clients current working directory  and  file with the same name on Server's side)\n";
+    }
+    else if(strcmp(commandAsked, "download") == 0)
+    {
+        return "download <file>\n"
+               "Request to receive <file> from Server's directory to client side.\n";
+    }
+    else if(strcmp(commandAsked, "archServer") == 0)
+    {
+        return "archServer <fileName>.tar\n"
+               "Using fork, exec and tar utilities create a child process that will collect all the files currently available on the the Server side and store them in the <filename>.tar archive.\n";
+    }
+    else if(strcmp(commandAsked, "killServer") == 0)
+    {
+        return "killServer\n"
+               "Sends a kill request to the Server\n";
+    }
+    else if(strcmp(commandAsked, "quit") == 0)
+    {
+        return "quit\n"
+               "Send write request to Server side log file";
+    }
+    else
+    {
+        return "Invalid command\n";
     }
 }
 
@@ -175,6 +207,10 @@ char* read_file(const char* filename, int line_number)
 
 int write_to_file(const char* filename, int line_number, const char* content) 
 {
+    printf("In write_to_file\n");
+    printf("Filename: %s\n", filename);
+    printf("Line number: %d\n", line_number);
+    printf("Content: %s\n", content);
     FILE* file = fopen(filename, "a+"); // Open file in append mode or create if not exists
     if (file == NULL) 
     {
@@ -272,6 +308,7 @@ int upload_file(const char* filename, const char* server_dir)
 
 int download_file(const char* filename, const char* server_dir) 
 {
+    printf("In download_file\n");
     // Check if the file exists in the server directory
     char server_filename[256];
     snprintf(server_filename, sizeof(server_filename), "%s/%s", server_dir, filename);
@@ -376,9 +413,25 @@ int archive_server(const char* filename, const char* server_dir)
     }
 }
 
-void handle_request(request_t request, int client_fd, queue_t *waiting_list, queue_t *connected_list, char *dirname, int max_clients)
+void handle_request(request_t request, queue_t *waiting_list, queue_t *connected_list, char *dirname, int max_clients)
 {
+    printf("In handle_request\n");
     char log[100];
+    int client_fd;
+    char client_fifo[CLIENT_FIFO_NAME_LEN];
+    printf("Client %d requested operation %d\n", request.client_pid, request.operation_type);
+
+    snprintf(client_fifo, CLIENT_FIFO_NAME_LEN, CLIENT_FIFO_TEMPLATE, request.client_pid);
+    if((client_fd = open(client_fifo, O_WRONLY)) == -1)
+    {
+        fprintf(stderr, "Error opening client FIFO: %s (errno=%d)\n", strerror(errno), errno);
+        exit(EXIT_FAILURE);
+    }
+    else
+    {
+        fprintf(stdout, "Client FIFO opened: %d\n", client_fd);
+    }
+
     char *file_list = NULL;
     char *file_content = NULL;
     int status = -2;
@@ -417,14 +470,13 @@ void handle_request(request_t request, int client_fd, queue_t *waiting_list, que
             send_response(SUCCESS, "Connected to server.\n", client_fd, request.client_pid);
         }
     }
-    else if(is_client_in_queue(connected_list, request.client_pid == 1))
-    {
+
         switch (request.operation_type)
         {
             case HELP:
                 if(strcmp(request.command.filename, "") != 0)
                 {
-                    send_response(SUCCESS, help_for_operation(request.operation_type), client_fd, request.client_pid);
+                    send_response(SUCCESS, help_for_operation(request.command.filename), client_fd, request.client_pid);
                 }
                 else
                 {
@@ -493,37 +545,15 @@ void handle_request(request_t request, int client_fd, queue_t *waiting_list, que
                         send_response(FAILURE, "Error archiving server. Child process terminated abnormally.\n", client_fd, request.client_pid);
                     }
                 break;
-            case KILL_SERVER:
-                while (!is_empty(connected_list)) 
-                {
-                    pthread_mutex_lock(&connected_list_mutex);
-                    int child_pid = dequeue(connected_list);
-                    pthread_mutex_unlock(&connected_list_mutex);
-                    kill(child_pid, SIGKILL);
-                }
-                while (!is_empty(waiting_list)) 
-                {
-                    pthread_mutex_lock(&waiting_list_mutex);
-                    int child_pid = dequeue(waiting_list);
-                    pthread_mutex_unlock(&waiting_list_mutex);
-                    kill(child_pid, SIGKILL);
-                }
-                kill(getpid(), SIGKILL);
-                // Write to log file
-                snprintf(log, sizeof(log), "Server killed. PID: %d\n", getpid());
-                log_message(log);
-                break;
             case QUIT:
-                // Kill the client process
-                kill(request.client_pid, SIGKILL);
-                // Write to log file
                 snprintf(log, sizeof(log), "Client %d killed.\n", request.client_pid);
                 log_message(log);
+                kill(request.client_pid, SIGKILL);
                 break;
             default:
                 break;
         }
-    }
+    
 }
 
 int main(int argc, char *argv[])
@@ -532,11 +562,14 @@ int main(int argc, char *argv[])
     int client_fd = 0;
     char server_fifo[100];
     request_t request; 
+    response_t connction_response;
     queue_t *waiting_list = create_queue();
     queue_t *connected_list = create_queue();
     char *dirname = NULL;
     char message[100];
     char client_fifo[CLIENT_FIFO_NAME_LEN];
+
+    signal(SIGINT, sigint_handler);
 
     if(argc != 3)
     {
@@ -579,14 +612,15 @@ int main(int argc, char *argv[])
     }
    
     umask(0);
-    snprintf(server_fifo, sizeof(server_fifo), SERVER_FIFO_TEMPLATE, getpid());
-    if(mkfifo(SERVER_FIFO_TEMPLATE, 0666) == -1)
+    snprintf(server_fifo, SERVER_FIFO_NAME_LEN, SERVER_FIFO_TEMPLATE, getpid());
+    print(server_fifo);
+    if(mkfifo(server_fifo, 0666) == -1)
     {
         fprintf(stderr, "Error creating server FIFO: %s (errno=%d)\n", strerror(errno), errno);
         exit(EXIT_FAILURE);
     }
 
-    server_fd = open(SERVER_FIFO_TEMPLATE, O_RDONLY|O_NONBLOCK);
+    server_fd = open(server_fifo, O_RDONLY|O_NONBLOCK);
     if(server_fd == -1)
     {
         fprintf(stderr, "Error opening server FIFO: %s (errno=%d)\n", strerror(errno), errno);
@@ -594,7 +628,7 @@ int main(int argc, char *argv[])
     }
 
     /* Open the server FIFO for writing so that it doesn't return EOF */
-    dummy_fd = open(SERVER_FIFO_TEMPLATE, O_WRONLY );
+    dummy_fd = open(server_fifo, O_WRONLY );
     if(dummy_fd == -1)
     {
         fprintf(stderr, "Error opening server FIFO for writing: %s (errno=%d)\n", strerror(errno), errno);
@@ -606,10 +640,16 @@ int main(int argc, char *argv[])
     log_message(message);
     fprintf(stdout, "Server started with PID=%d\n", getpid());
 
-    pthread_mutex_lock(&global_var_mutex);
-    while(!kill_signal_received)
+    while(1)
     {
-
+        if(request.operation_type == KILL_SERVER)
+        {
+            printf("Kill server request received\n");
+            // Directly initiate shutdown sequence
+            unlink(server_fifo); // Remove the server FIFO to clean up
+            exit(EXIT_SUCCESS);  // Exit directly
+        }
+        /*
         // Check if connected clients have plots available
         if (!is_empty(connected_list) && !is_empty(waiting_list))
         {
@@ -636,10 +676,11 @@ int main(int argc, char *argv[])
                 }
             }
         }
-
+        */
         ssize_t bytes_read;
         while ((bytes_read = read(server_fd, &request, sizeof(request_t))) == -1) 
         {
+
             if (errno == EINTR) {
                 // Retry the read operation if interrupted by a signal
                 continue;
@@ -653,50 +694,87 @@ int main(int argc, char *argv[])
                 exit(EXIT_FAILURE);
             }
         }
-
-        pid_t pid = fork();
-        if(pid < 0)
+        printf("Operation type: %d\n", request.operation_type);
+        if(request.operation_type == KILL_SERVER)
         {
-            perror("Fork failed");
-            exit(EXIT_FAILURE);
+            printf("Kill server request received\n");
+            // Directly initiate shutdown sequence
+            unlink(server_fifo); // Remove the server FIFO to clean up
+            exit(EXIT_SUCCESS);  // Exit directly
         }
-        else if(pid == 0)
-        {
-            // Child process
-            client_fd = open(client_fifo, O_WRONLY);
-            if(client_fd == -1)
+
+            pid_t pid = fork();
+            if(pid < 0)
             {
-                fprintf(stderr, "Error opening client FIFO: %s (errno=%d)\n", strerror(errno), errno);
+                perror("Fork failed");
                 exit(EXIT_FAILURE);
             }
-
-            handle_request(request, client_fd, waiting_list, connected_list, dirname, max_clients);
-
-            exit(EXIT_SUCCESS);
-        }
-        else 
-        {
-            //Parent process
-            int status;
-            waitpid(pid, &status, 0);
-            if(WIFEXITED(status))
+            else if(pid == 0)
             {
-                int exit_status = WEXITSTATUS(status);
-                if(exit_status == 0)
+                if(request.operation_type == NONE)
                 {
-                    fprintf(stdout, "Child process %d exited successfully\n", pid);
+                    enqueue(waiting_list, request.client_pid);
+                    int client_fd2;
+                    char client_fifo2[CLIENT_FIFO_NAME_LEN];
+                    snprintf(client_fifo2, CLIENT_FIFO_NAME_LEN, CLIENT_FIFO_TEMPLATE, request.client_pid);
+                    if((client_fd2 = open(client_fifo2, O_WRONLY)) == -1)
+                    {
+                        fprintf(stderr, "Error opening client FIFO: %s (errno=%d)\n", strerror(errno), errno);
+                        exit(EXIT_FAILURE);
+                    }
+                    else
+                    {
+                        fprintf(stdout, "Client FIFO2 opened: %d\n", client_fd2);
+                    }
+                    if(kill_signal_received == 1)
+                    {
+                        exit(EXIT_SUCCESS);
+                    }
+                    send_response(SUCCESS, "Connecteeeeeed.\n", client_fd2, request.client_pid);
+                }
+                else if(request.operation_type == KILL_SERVER)
+                {
+                    printf("Kill server request received\n");
+                    kill_signal_received = 1;
+                    printf("Kill signal value: %d\n", kill_signal_received);
+                    kill_parent_and_children(getppid());
+                    exit(EXIT_SUCCESS);
                 }
                 else
                 {
-                    fprintf(stderr, "Child process %d exited with status %d\n",pid, exit_status);
+                    handle_request(request, waiting_list, connected_list, dirname, max_clients);
+                    exit(EXIT_SUCCESS);
                 }
+    
             }
-            else
+            else 
             {
-                fprintf(stderr, "Child process %d terminated abnormally\n", pid);
+                if(kill_signal_received==1)
+                {
+                    kill_parent_and_children(getpid());
+                }
+                //Parent process
+                int status;
+                waitpid(pid, &status, 0);
+                if(WIFEXITED(status))
+                {
+                    int exit_status = WEXITSTATUS(status);
+                    if(exit_status == 0)
+                    {
+                        fprintf(stdout, "Child process %d exited successfully\n", pid);
+                    }
+                    else
+                    {
+                        fprintf(stderr, "Child process %d exited with status %d\n",pid, exit_status);
+                    }
+                }
+                else
+                {
+                    fprintf(stderr, "Child process %d terminated abnormally\n", pid);
+                }
+                
             }
-        }
+        
     }
-    pthread_mutex_unlock(&global_var_mutex);
     exit(EXIT_SUCCESS);
 }
